Added initial support for streaming out PDFs directly to the disc. This feature is used by the LargeTest test program. Streaming out directly to disc is currently supported using PdfDocument and PdfImmediateWriter. Currently one can create broken documents very easy using this combination. A class PdfStreamedDocument will be added instead to support streaming out to disc. PdfDocument will be used for PDF file manipulations in memory and PdfStreamedDocument should be used for the sole purpose of creating new PDF files.
PdfImage and PdfFont have yet to be adapted to make use of streaming. They still allocate memory for images and fonts which is not 
needed. Additionally the filters have to be modified to support streaming, too. Documents created by PdfImmediateWriter are not 
compressed yet.

My next step will be to create a useful public API for this feature using PdfStreamedDocument which won't allow to create broken 
PDFs.



git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@392 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5