On GCC 3.4 & above, use the ::__gnu_cxx::__pool_alloc<T> allocator to automatically pool allocations of TRefCountedBuffer objects. This is not a performance win - in testing, it shows little impact either way on real world performance. Again, however, it'll help us reduce memory fragmentation with this heavily used class.
It is trivial to try different custom allocators or to disable the use of the
custom allocator.

Other platforms and older versions of gcc continue to use std::allocator via
operator new()'s default behaviour.

This change is binary incompatible but source compatible, and existing users
won't notice anything different.


git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@728 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5