FIXED: PdfStreamedDocument should create correct XREF tables now.
There were a semantic misunderstanding of mine in PdfOutputDevice. PdfOutputDevice::GetLength counts the bytes written to the device. If you seek 
back this won't be decreased. As a reason I added PdfOutputDevice::Tell which returns the current offset from the beginning of the file. Please make 
sure you use always the correct method. Most of the time Tell is the correct choice over GetLength.



git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@799 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5