Finish splitting compiler compat stuff into private and public parts.
- Remove all access to PdfCompilerCompatPrivate.h from PoDoFo's headers
- Introduce PdfDefinesPrivate.h include in _EVERY_ .cpp files
- Have the new PdfDefinesPrivate.h include PdfCompilerCompatPrivate.h, so that's visible
  in every .cpp file but not in the headers

It is HIGLY LIKELY THAT THIS CHANGE BREAKS OTHER PLATFORMS AND COMPILERS. Work will
be required to get them building again due to type visibility issues, header include
ordering problems, etc.

The end result will be much cleaner, easier-to-use headers that'll be less
likely to interfere with the library client's sources and the operation of the
library client, though.



git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@1107 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5