Introduce PdfMemDocument::IsLoaded()
When an encrypted file fails to be loaded due to incorrect
or none password provided, then the PdfMemDocument is left
in a state which can lead to a crash (when for example
trying to access AcroForm). The added IsLoaded() is to check
whether the load was properly completed.

As a side-effect, a bug with not freed (but not leaked) parser
at PdfMemDocument::SetPassword() had been fixed.


git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@1656 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5