Patch by Francesco Pretto: Fix clamping (avoiding crashes) in PdfPagesTree::InsertPage(const PdfRect&, int)
This patch is from the podofo-users mailing list post specified below and fixes
a crash when called with int index 0, in PdfPagesTreeCache::AddPageObject(), by
array out-of-bounds access. The new clamping allows an index equal to number of
pages but will clamp negative input indices to 0 instead of the page count - 1,
positive input indices to the page count (meaning "insert after the last page")
which avoids the call to the latter method with a negative index (crash cause).

Metadata of the patch-source mailing list post:
Date: Fri, 23 Feb 2018 19:27:40 +0100
Message-ID: <CALas-igwWbEOr7dW1=t0c6N20Y4hNuD13seSPgaVB_0KUJ+zQg@mail.gmail.com>
Subject prefix: [PATCH 05/13],
further Subject: PdfPagesTree::InsertPage: Fix "atIndex" is really a 0-based index


git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@1898 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5