Fix CVE-2018-5783 by introducing singleton limit for indirect objects
The limit is the standard one of 8,388,607 but can be changed by
a newly introduced inline method (also a getter is provided).
This also introduced a new define PODOFO_SIZE_FORMAT for formatting
size_t values portably in printf()-like methods of PdfError etc.
Please also cf. issue #4 in the issue tracker.


git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@1949 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5